requirements.txt
pandas
numpy
scikit-learn
joblib
Flask
web3
python-dotenv